feat(zq): configure multiple init datas for test mode
Test Mode now allows you to customize test-only init data. You can make
multiple init datas. They are saved to your `zquest.cfg`. A basic name
is auto-assigned, but you can set your own name by modifying your
config.

When you create a new test init data, your current quest's init data is
used as the starting point. Later modifying the quest init data has no
impact on your test init datas.

You can also directly modify the init data values in the config, though
the format is not documented anywhere.
