+++ src/dialog/test_quest_dlg.h
@@ -15,7 +15,7 @@ void call_testqst_dialog();
 class TestQstDialog: public GUI::Dialog<TestQstDialog>
 {
 public:
-	enum class message { REFR_INFO, OK, CANCEL };
+	enum class message { REFR_INFO, CREATE_INIT_DATA, EDIT_INIT_DATA, OK, CANCEL };
 
 	TestQstDialog();
