+++ src/init.h
@@ -12,6 +12,10 @@
 
 #include "tab_ctl.h"
 #include "base/zc_alleg.h"
+#include <string>
+
+class gamedata;
+class zinitdata;
 
 #define MAXINITTABS 5
 
@@ -36,5 +40,8 @@ int32_t d_bombratio_proc(int32_t msg,DIALOG *d,int32_t c);
 extern TABPANEL init_tabs[];
 int32_t doInit(zinitdata *zinit, bool isZC);
 void resetItems(gamedata *data, zinitdata *zinit, bool freshquest);
+
+std::string serialize_init_data_delta(zinitdata *base, zinitdata *changed);
+zinitdata *apply_init_data_delta(zinitdata *base, std::string delta);
 #endif
