+++ src/zc/zc_init.h
@@ -16,7 +16,5 @@
 int32_t onCheatConsole();
 void center_zc_init_dialog();
 zinitdata *copyIntoZinit(gamedata *src);
-std::string serialize_init_data_delta(zinitdata *base, zinitdata *changed);
-zinitdata *apply_init_data_delta(zinitdata *base, std::string delta);
 #endif
